+++ src/MVClip.h
@@ -4,15 +4,15 @@
 
 class MVClipDicks : public MVAnalysisData {
 	int32_t nBlkCount;
-	float nSCD1;
-	float nSCD2;
+	double nSCD1;
+	double nSCD2;
 	const VSAPI *vsapi;
 public:
-	MVClipDicks(VSNodeRef *vectors, float nSCD1, float nSCD2, const VSAPI *_vsapi);
+	MVClipDicks(VSNodeRef *vectors, double nSCD1, double nSCD2, const VSAPI *_vsapi);
 	~MVClipDicks();
 	inline int32_t GetBlkCount() const { return nBlkCount; }
-	inline float GetThSCD1() const { return nSCD1; }
-	inline float GetThSCD2() const { return nSCD2; }
+	inline double GetThSCD1() const { return nSCD1; }
+	inline double GetThSCD2() const { return nSCD2; }
 };
 
 class MVClipBalls : public FakeGroupOfPlanes {
@@ -24,7 +24,7 @@ public:
 	void Update(const VSFrameRef *fn);
 	inline const FakeBlockData& GetBlock(int32_t nLevel, int32_t nBlk) const { return GetPlane(nLevel)[nBlk]; }
 	bool IsUsable() const;
-	bool IsSceneChange(float nSCD1, float nSCD2) const { return FakeGroupOfPlanes::IsSceneChange(nSCD1, nSCD2); }
+	bool IsSceneChange(double nSCD1, double nSCD2) const { return FakeGroupOfPlanes::IsSceneChange(nSCD1, nSCD2); }
 };
 
 #endif
